refactor: FeedbackDialog から FeedbackImagePicker を分離 (#266)
